Refactor much of random logic out of RomHandler
- Some edits are difficult to make because many methods within the
RomHandler do randomization on their own, so I have to untangle a lot
of code in order to make the changes I want to make. So I'm moving as
much logic that generates randomness outside of the RomHandler, and the
RomHandler will ideally only be used as an abstracted interface with
the Rom.