Update thematic wild randomization
- I've tried to make the randomizer a little smarter when determining
what pokemon to replace the OG pokemon with when using vanilla randomization.
Originally the logic only used the base pokemon's primary type to determine
the possible choices for replacement, falling back to the secondary type
if not enough options were found.

Now, the randomizer checks what the most common type in an area is. If
any of the OG pokemon's types match what the most common type is in that
area, then choose a replacement that matches the common type. Otherwise
default back to using the OG pokemon's primary type as the replacement.

The motivation
behind this change was that ICE pokemon in GenIV were not being reliably
being replaced with other ICE types, since most ICE types of the gen have
a different primary type. So an area that used to be mostly ICE pokemon became
a random assortment of other types, and I felt that created a tonal mismatch.